updated App.js file with routes to pages

+++ src/App.js
@@ -1,33 +1,69 @@
 import { useState, useEffect } from 'react'
+import AuthPage from './pages/AuthPage/AuthPage'
+import HomePage from './pages/HomePage/HomePage'
+import { Route, Routes } from 'react-router-dom'
 import BookmarkList from './components/BookmarkList/BookmarkList'
 import styles from './App.module.scss'
 
-
 export default function App(){
-    const [allBookmarks, setBookmarks] = useState([])
-    const [newBookmark, setNewBookmark] = useState({
-        title: '',
-        url: ''
-    })
-
-    //createBookmark
-    const createBookmark = async () => {
-        const body = {...newBookmark}
+    const [user, setUser] = useState(null)
+    const [token, setToken] = useState('')
+ 
+    const signUp = async (credentials) => {
         try {
-            const response = await fetch('/api/bookmarks', {
+            const response = await fetch('/api/user', {
                 method: 'POST',
                 headers: {
                     'Content-Type': 'application/json'
                 },
-                body: JSON.stringify(body)
+                body: JSON.stringify(credentials)
             })
-            const createdBookmark = await response.json()
-            const bookmarksCopy = [createdBookmark,...allBookmarks]
-            setBookmarks(bookmarksCopy)
-            setNewBookmark({
-                title: '',
-                url: ''
+            const data = await response.json()
+            setUser(data.user)
+            localStorage.setItem('token', data.token)
+            localStorage.setItem('user', JSON.stringify(data.user))
+        } catch (error) {
+            console.error(error)
+        }
+    }
+
+    const login = async (credentials) => {
+        try {
+            const response = await fetch('/api/user/login', {
+                method: 'POST',
+                headers: {
+                    'Content-Type': 'application/json'
+                },
+                body: JSON.stringify(credentials)
             })
+            const data = await response.json()
+            const tokenData = data.token
+            localStorage.setItem('token', tokenData)
+            setToken(tokenData)
+            const userData = data.user
+            localStorage.setItem('user', JSON.stringify(userData))
+            setUser(userData)
+        } catch (error) {
+            console.error(error)
+        }
+    }
+
+    //createBookmark
+    const createBookmark = async (blogData, token) => {
+        if (!token) {
+            return
+        }
+        try {
+            const response = await fetch('/api/bookmarks', {
+                method: 'POST',
+                headers: {
+                    'Content-Type': 'application/json',
+                    'Authorization': `Bearer ${token}`
+                },
+                body: JSON.stringify(blogData)
+            })
+            const data = await response.json()
+            return data
         } catch (error) {   
             console.error(error)
         }
@@ -35,20 +71,22 @@ export default function App(){
 
     //updateBookmark
 
-    const updateBookmark = async (id, subject) => {
+    const updateBookmark = async (newBlogData, id, token) => {
+        if (!token) {
+            return
+        }
         try {
             const response = await fetch(`/api/bookmarks/${id}`, {
                 method: 'PUT',
                 headers: {
-                    'Content-Type': 'application/json'
+                    'Content-Type': 'application/json',
+                    'Authorization': `Bearer ${token}`
                 },
-                body: JSON.stringify(subject)
+                body: JSON.stringify(newBlogData)
             })
-            const index = allBookmarks.findIndex((bookmark) => id === bookmark._id)
-            const allBookmarksCopy = [...allBookmarks]
+            
             const data = await response.json()
-            allBookmarksCopy[index] = {...allBookmarksCopy[index], ...subject}
-            setBookmarks(allBookmarksCopy)
+            return data
         } catch (error) {
             console.error(error)
         }
@@ -56,50 +94,80 @@ export default function App(){
 
 
     //deleteBookmark
-    const deleteBookmark = async (id) => {
+    const deleteBookmark = async (id, token) => {
+        if (!token) {
+            return
+        }
         try {
-            const index = allBookmarks.findIndex((bookmark) => bookmark._id === id)
-            const allBookmarksCopy = [...allBookmarks]
             const response = await fetch(`/api/bookmarks/${id}`, {
                 method: 'DELETE',
                 headers: {
-                    'Content-Type': 'application/json'
+                    'Authorization': `Bearer ${token}`
                 }
             })
-            await response.json()
-            allBookmarksCopy.splice(index, 1)
-            setBookmarks(allBookmarksCopy)
+            const data = await response.json()
+            return data
         } catch (error) {
             console.error(error)
         }
     }
+
     //getBookmarks
     const getBookmarks = async () => {
         try{
             const response = await fetch('/api/bookmarks')
-            const foundBookmarks = await response.json()
-            setBookmarks(foundBookmarks.reverse())
+            const data = await response.json()
+            return data
         } catch(error){
             console.error(error)
         }
     }
-    useEffect(() => {
-        getBookmarks()
-    }, [])
+    // useEffect(() => {
+    //     getBookmarks()
+    // }, [])
     return(
-        <>
-			
+
+        <div className={styles.App}>
             <div className={styles.banner}>
                 <h1>Bookmarks</h1>
             </div>
-            <BookmarkList
-            newBookmark={newBookmark}
-            setNewBookmark={setNewBookmark}
-            createBookmark={createBookmark}
-            updateBookmark={updateBookmark}
-            allBookmarks={allBookmarks}
-            deleteBookmark={deleteBookmark}
-            />
-        </>
+            <Routes>
+                <Route path="/" element={
+                <HomePage
+                    user={user}
+                    token={token}
+                    setToken={setToken}
+                    setUser={setUser}
+                    getBookmarks={getBookmarks}
+                    createBookmark={createBookmark}
+                    deleteBookmark={deleteBookmark}
+                    updateBookmark={updateBookmark}
+                />}></Route>
+                <Route path="/register" element={
+                <AuthPage
+                    setUser={setUser}
+                    setToken={setToken}
+                    signUp={signUp}
+                    login={login}
+                />}></Route>
+            </Routes>
+        </div>
+
+
+
+        // <>
+			
+        //     <div className={styles.banner}>
+        //         <h1>Bookmarks</h1>
+        //     </div>
+        //     <BookmarkList
+        //     newBookmark={newBookmark}
+        //     setNewBookmark={setNewBookmark}
+        //     createBookmark={createBookmark}
+        //     updateBookmark={updateBookmark}
+        //     allBookmarks={allBookmarks}
+        //     deleteBookmark={deleteBookmark}
+        //     />
+        // </>
     )
 }
